worker/requirements.txt
tweepy
celery
nose
python-dateutil
pytz
https://github.com/dryan/twitter-text-py/archive/master.zip
